LocalConnections.cs split into LocalConnectionToClient/Server.cs
